+++ packages/provider/src/index.ts
@@ -1,3 +1,4 @@
 export * from './embedding-model/index';
 export * from './errors/index';
+export * from './json-value/index';
 export * from './language-model/index';