+++ packages/provider/src/language-model/v1/language-model-v1-provider-metadata.ts
@@ -1,0 +1,25 @@
+import { JSONValue } from '../../json-value/json-value';
+
+/**
+ * Additional provider-specific metadata. They are passed through
+ * to the provider from the AI SDK and enable provider-specific
+ * functionality that can be fully encapsulated in the provider.
+ *
+ * This enables us to quickly ship provider-specific functionality
+ * without affecting the core AI SDK.
+ *
+ * The outer record is keyed by the provider name, and the inner
+ * record is keyed by the provider-specific metadata key.
+ *
+ * ```ts
+ * {
+ *   "anthropic": {
+ *     "cacheControl": { "type": "ephemeral" }
+ *   }
+ * }
+ * ```
+ */
+export type LanguageModelV1ProviderMetadata = Record<
+  string,
+  Record<string, JSONValue>
+>;